requirements.txt
Pillow
django-simple-history
django-crispy-forms
crispy-bootstrap5